move: impl/recipe/filter/CVanillaRecipe -> impl/recipe/CVanillaRecipe
